+++ kernel/timer/timer.c
@@ -3,33 +3,33 @@
 #include "builddef.h"
 #include "drivers/pit.h"
 #include "interrupt.h"
+#include "stdint.h"
 #include "tasks.h"
-#include "utils/debug.h"
 
 #define TICK_FREQ_HZ 100
 
 private
-unsigned int _ticks;
+size_t _ticks;
 
 // 0x20 - Programmable Interval Timer
 private
-void ISR_SYSTIME(unsigned int* return_reg, struct cpu_state* unused) {
+void ISR_SYSTIME(size_t* return_reg, struct cpu_state* unused) {
     _ticks++;
     task_isr_priority();
 }
 
 public
-unsigned int getticks() { return _ticks; }
+size_t getticks() { return _ticks; }
 
 public
-void delay(unsigned int ms) {
-    unsigned int stop = _ticks * 10 + ms;
+void delay(size_t ms) {
+    size_t stop = _ticks * 10 + ms;
     while (_ticks * 10 < stop) {
     }
 }
 
 public
-int timer_init() {
+int32_t timer_init() {
     pit_setfreq(TICK_FREQ_HZ);
     interrupt_register(INT_SYSTIME, ISR_SYSTIME);
     return 0;
